Major update to API, combining the two rest controllers as well as adding functionality to get a specific users servers. Update to frontend to reflect API changes as well as checking administrator privileges.
