--- src/main/java/com/example/webServer/web/rest/Map.java
@@ -1,4 +1,0 @@
-package com.example.webServer.web.rest;
-
-public interface Map<T, T1> {
-}
