+++ src/main/java/com/example/webServer/web/rest/ServerRestController.java
@@ -1,5 +1,6 @@
 package com.example.webServer.web.rest;
 
+import com.example.webServer.data.repositories.Accounts;
 import com.example.webServer.services.ServerService;
 import com.example.webServer.web.errors.BadRequestException;
 import com.example.webServer.web.models.Server;
@@ -10,50 +11,66 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/servers")
+@RequestMapping(
+        value = "/api",
+        method = {RequestMethod.GET, RequestMethod.PUT, RequestMethod.POST}
+)
 public class ServerRestController {
     private final ServerService serverService;
+    private final Accounts accounts = new Accounts();
 
     public ServerRestController(ServerService serverService) {
         this.serverService = serverService;
     }
 
-    @GetMapping
+    @GetMapping("/servers")
     public List<Server> getAll(@RequestParam(name="serverName", required = false)String serverName){
         return this.serverService.getAlLServers(serverName);
     }
 
-    @GetMapping("/getUserServers")
+    @GetMapping("/servers/getUserServers")
     public List<Server> getUserServers(@RequestBody String json){
-        String[] accountInformation = parseLoginJson(json);
-        String username = accountInformation[0].replaceAll("\"", "");
+        String username = json.replaceAll("\"", "");
         return this.serverService.getUserServers(username);
     }
 
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public Server createServer(@RequestBody Server server){
-        return this.serverService.createOrUpdate(server);
-    }
+    // setting default test username/password
+    @GetMapping("/login")
+    @CrossOrigin()
+    public String getLoginToken(@RequestBody String jsonLoginInformation){
+        String[] loginInformation = parseLoginJson(jsonLoginInformation);
 
-    @GetMapping("/{id}")
-    public Server getServer(@PathVariable("id")long id){
-        return this.serverService.getServer(id);
-    }
+        String username = loginInformation[0].replaceAll("\"", "");
+        String password = loginInformation[1].replaceAll("\"", "");
 
-    @PutMapping("/{id}")
-    public Server updateServer(@PathVariable("id")long id, @RequestBody Server server){
-        if(id != server.getServerId()){
-            throw new BadRequestException("ids do not match");
+        if (!accounts.isValidAccount(username, password)){
+            throw new BadRequestException("invalid login information");
+        } else {
+            // TODO give an actual token
+            return "{\"token\": \"" +username+"\"}";
         }
-        return this.serverService.createOrUpdate(server);
     }
 
-    @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.RESET_CONTENT)
-    public void deleteServer(@PathVariable("id")long id){
-        this.serverService.deleteServer(id);
+    @GetMapping("/createAccount")
+    @CrossOrigin()
+    public String createNewAccount(@RequestBody String json){
+        String[] accountInformation = parseLoginJson(json);
+        String username = accountInformation[0].replaceAll("\"", "");
+        String password = accountInformation[1].replaceAll("\"", "");
+        System.out.println(json);
+
+        if(accounts.accountExists(username)){
+            throw new BadRequestException("account already exists");
+        } else {
+            if (accounts.createAccount(username, password)) {
+                return "{\"token\": \"" +username+"\"}";
+            } else {
+                return "{\"accountCreationStatus\": \"failed\"}";
+            }
+        }
+
     }
+
 
     private String[] parseLoginJson(String jsonObj){
         String[] loginInformation = jsonObj.split(",");
